Moved functionality from aspect-configuration-component to service
