+++ frontend/src/app/modules/aspect-configuration/aspect-configuration.component.ts
@@ -2,16 +2,10 @@ import {Component, OnInit} from '@angular/core';
 import {ApplicationService} from "../../services/application.service";
 import {ActivatedRoute, ParamMap} from "@angular/router";
 import {Page} from "../../models/page.model";
-import {combineLatest, Observable} from "rxjs";
-import {
-  ExtendedPerformanceAspect,
-  PerformanceAspect,
-  PerformanceAspectType
-} from "../../models/perfomance-aspect.model";
+import {Observable} from "rxjs";
+import {ExtendedPerformanceAspect, PerformanceAspectType} from "../../models/perfomance-aspect.model";
 import {Application} from "../../models/application.model";
 import {AspectConfigurationService} from "./services/aspect-configuration.service";
-import {BrowserInfoDto} from "../../models/browser.model";
-import {map} from "rxjs/operators";
 
 @Component({
   selector: 'osm-aspect-configuration',
@@ -29,78 +23,18 @@ export class AspectConfigurationComponent implements OnInit {
   constructor(private route: ActivatedRoute, private applicationService: ApplicationService, private aspectConfService: AspectConfigurationService) {
     this.application$ = applicationService.selectedApplication$;
     this.page$ = applicationService.selectedPage$;
-    this.prepareExtensionOfAspects();
+    this.performanceAspects$ = aspectConfService.performanceAspects$;
+    this.aspectTypes$ = this.aspectConfService.uniqueAspectTypes$;
+    this.aspectConfService.prepareExtensionOfAspects();
   }
 
   ngOnInit() {
-    this.getBrowserInfos();
+    this.aspectConfService.loadBrowserInfos();
     this.route.paramMap.subscribe((params: ParamMap) => {
-      this.getApplication(params.get('applicationId'));
-      this.getPage(params.get('pageId'));
+      this.aspectConfService.loadApplication(params.get('applicationId'));
+      this.aspectConfService.loadPage(params.get('pageId'));
     });
-    this.initAspectTypes();
+    this.aspectConfService.initAspectTypes();
   }
 
-  initAspectTypes() {
-    this.aspectTypes$ = this.performanceAspects$.pipe(
-      map((extendedAspects: ExtendedPerformanceAspect[]) => {
-        const uniqueAspectTypes = [];
-        const lookupMap = new Map();
-        for (const aspect of extendedAspects) {
-          if (!lookupMap.has(aspect.performanceAspectType.name)) {
-            lookupMap.set(aspect.performanceAspectType.name, true);
-            uniqueAspectTypes.push(aspect.performanceAspectType)
-          }
-        }
-        return uniqueAspectTypes
-      })
-    )
-  }
-
-  private prepareExtensionOfAspects() {
-    this.performanceAspects$ = combineLatest(this.applicationService.performanceAspectsForPage$, this.aspectConfService.browserInfos$).pipe(
-      map(([aspects, browserInfos]: [PerformanceAspect[], BrowserInfoDto[]]) => {
-        const extendedAspects: ExtendedPerformanceAspect[] = this.extendAspects(aspects, browserInfos);
-        return extendedAspects;
-      })
-    )
-  }
-
-  extendAspects(aspects: PerformanceAspect[], browserInfos: BrowserInfoDto[]): ExtendedPerformanceAspect[] {
-    const extendedAspects: ExtendedPerformanceAspect[] = [];
-    if (aspects.length > 0 && browserInfos.length > 0) {
-      aspects.forEach((aspect: PerformanceAspect) => {
-        const additionalInfos = browserInfos.filter((browserInfo: BrowserInfoDto) => browserInfo.browserId == aspect.browserId);
-        let extension: BrowserInfoDto;
-        if (additionalInfos.length == 1) {
-          extension = additionalInfos[0];
-        } else {
-          extension = {
-            browserId: aspect.browserId,
-            browserName: 'Unknown',
-            operatingSystem: 'Unknown',
-            deviceType: {name: 'Unknown', icon: 'question'}
-          }
-        }
-        extendedAspects.push({...aspect, ...extension})
-      });
-    }
-    return extendedAspects;
-  }
-
-  getApplication(applicationId: string) {
-    this.aspectConfService.loadApplication(applicationId).subscribe((app: Application) => {
-      this.applicationService.selectedApplication$.next(app);
-    })
-  }
-
-  getPage(pageId: string) {
-    this.aspectConfService.loadPage(pageId).subscribe((page: Page) => {
-      this.applicationService.selectedPage$.next(page);
-    })
-  }
-
-  getBrowserInfos() {
-    this.aspectConfService.loadBrowserInfos()
-  }
 }
